pyproject.toml
[project]
name = "confluence-markdown"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4 (>=4.13, <5.0)",
    "html2text (>=2025.4, <2026.0)",
    "pip (>=25.1, <26.0)",
    "python-dotenv (>=1.1, <2.0)",
    "requests (>=2.32, <3.0)",
]
